Fix precommit windows warnings

+++ modules/gapi/test/common/gapi_imgproc_tests_inl.hpp
@@ -561,8 +561,8 @@ TEST_P(BoundingRectVectorTest, AccuracyTest)
     {
         std::vector<cv::Point2f> in_vectorF(sz.width);
         for (int i = 0; i < sz.width; i++)
-            in_vectorF.push_back(cv::Point2f(exp(rng.uniform(-1, 6) * 3.0 * CV_LOG2),
-                                             exp(rng.uniform(-1, 6) * 3.0 * CV_LOG2)));
+            in_vectorF.push_back(cv::Point2f(expf(rng.uniform(-1, 6) * 3.0 * CV_LOG2),
+                                             expf(rng.uniform(-1, 6) * 3.0 * CV_LOG2)));
 
         // G-API code //////////////////////////////////////////////////////////////
         cv::GArray<cv::Point2f> in;
